requirements.txt
fastapi
kafka-python
nltk
pymongo
python-dotenv
requests
textblob
uvicorn
